requirements.txt
apache-airflow
apache-airflow-providers-apache-spark 
pyspark
python-telegram-bot